fix test_ordering_satisfy
